added angular material and more css

+++ src/app/modules/account/account-creation/account-creation.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { Router } from '@angular/router';
+import { ErrorPasswordComponent } from 'src/app/core/components/snack-bar/account-creation-confirm/error-password/error-password.component';
 import { HttpUserService } from 'src/app/core/http/user/httpUser.service';
 import { User } from 'src/app/core/model/user';
 import { AuthService } from 'src/app/core/services/auth.service';
@@ -14,7 +16,9 @@ export class AccountCreationComponent implements OnInit {
 
   creationForm: FormGroup;
 
-  constructor(private httpUser: HttpUserService, private router: Router, private auth: AuthService) { }
+  durationInSeconds: number = 5;
+
+  constructor(private httpUser: HttpUserService, private router: Router, private auth: AuthService, private _snackBar: MatSnackBar) { }
 
   ngOnInit(): void {
     this.creationForm = this.createNewAccount();
@@ -36,39 +40,51 @@ export class AccountCreationComponent implements OnInit {
     });
   }
 
+  get f() {
+    return this.creationForm.value;
+  }
+
   onCreate() {
-    const userToCreate: User = {
-      genre: this.creationForm.value.genre,
-      lastName: this.creationForm.value.lastName,
-      firstName: this.creationForm.value.firstName,
-      email: this.creationForm.value.email,
-      mobile: this.creationForm.value.mobile,
-      phone: this.creationForm.value.phone,
-      adress: this.creationForm.value.adress,
-      zipCode: this.creationForm.value.zipCode,
-      city: this.creationForm.value.city,
-      password: this.creationForm.value.password,
-      adminLevel: 'citoyen',
+    if (this.f.password === this.f.confirm) {
+      const userToCreate: User = {
+        genre: this.f.genre,
+        lastName: this.f.lastName,
+        firstName: this.f.firstName,
+        email: this.f.email,
+        mobile: this.f.mobile,
+        phone: this.f.phone,
+        adress: this.f.adress,
+        zipCode: this.f.zipCode,
+        city: this.f.city,
+        password: this.f.password,
+        adminLevel: 'citoyen',
+      }
+      this.httpUser.createUser(userToCreate).subscribe((account: any) => {
+        const payload = {
+          username: account.email,
+          password: this.creationForm.value.password
+        };
+        this.httpUser.userLogin(payload).subscribe((data: any) => {
+          this.auth.notifyObservable(data.accessToken);
+          this.auth.dataFromObservable.subscribe((authToken: string) => {
+            this.auth.authToken = authToken;
+            localStorage.setItem('token', authToken);
+            this.auth.notifyUserObservable(payload.username);
+            this.httpUser.getSingleUser(payload.username).subscribe((userId: string) => {
+              localStorage.setItem('userId', userId);
+              this.router.navigate(['/accueil']);
+            });
+          });
+        })
+      });
+    } else {
+      this._snackBar.openFromComponent(ErrorPasswordComponent, {
+        duration: this.durationInSeconds * 1000,
+        panelClass: "list-group-item-danger",
+        verticalPosition: "top",
+      });
     }
 
-    this.httpUser.createUser(userToCreate).subscribe((account: any) => {
-      const payload = {
-        username: account.email,
-        password: this.creationForm.value.password
-      };
-      this.httpUser.userLogin(payload).subscribe((data: any) => {
-        this.auth.notifyObservable(data.accessToken);
-        this.auth.dataFromObservable.subscribe((authToken: string) => {
-          this.auth.authToken = authToken;
-          localStorage.setItem('token', authToken);
-          this.auth.notifyUserObservable(payload.username);
-          this.httpUser.getSingleUser(payload.username).subscribe((userId: string) => {
-            localStorage.setItem('userId', userId);
-            this.router.navigate(['/accueil']);
-          });
-        });
-      })
-    });
   }
 
 }
